refactor: split org writer shared helpers and console/json writers
Extract common.py (normalization/validation helpers), trending.py
(trending table renderers and drift helpers), console.py (console
org-report renderers), and json.py (JSON builder/writer) from the
monolithic org/writers/__init__.py. The package root re-exports all
symbols to preserve existing import paths. Contract tests verify
both submodule and package-root importability.
